refactor(#1117): extract mega-functions into dedicated concern modules
Behavior-preserving follow-up to the directory split. Moves the three
critical-path mega-functions and the windowed streaming driver into their
own modules so no production module half exceeds ~1300 lines:

- parse_cell_value_schema_order (~1140 lines) -> cell_value.rs
- parse_raw_type_value (~1130 lines)          -> raw_type_value.rs
- parse_block_emit_windowed + parse_block_with_timestamps* +
  parse_one_partition_with_timestamps (sliding-window/timestamps driver,
  issue #827)                                 -> block_emit_windowed.rs

Pure relocation of whole impl methods (same impl V5CompressedLegacyParser,
same visibility); no logic change. row_data.rs 1902->739, raw_value.rs
1615->469 (prod), block_emit.rs 1534->614 prod. Refs epic #1116.
